student_management/v1/serializers/course_serializer.py
from rest_framework import serializers

from student_management.models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = "__all__"
